+++ src/filters/LessThanFilter.php
@@ -7,7 +7,7 @@
 namespace bitmatrix\omnisearch\filters;
 
 
-use craft\db\Query;
+use yii\db\Query;
 
 class LessThanFilter extends OmniSearchFilter
 {