+++ src/filters/LessThanOrEqualFilter.php
@@ -7,7 +7,7 @@
 namespace bitmatrix\omnisearch\filters;
 
 
-use craft\db\Query;
+use yii\db\Query;
 
 class LessThanOrEqualFilter extends OmniSearchFilter
 {